+++ src/components/forms/selectfield/SelectField.svelte
@@ -1,30 +1,45 @@
-<svelte:options customElement="khao-ui-selectfield" />
+<svelte:options customElement="khao-select-field" />
 
 <script lang="ts">
-
   export let label: string;
-  export let selectedValue: string;
+  export let selectedValue: string = "";
   export let options: string;
-  export let id: string = `khao-ui-selectfield-${label}`;
+  export let id: string = `khao-select-field-${label}`;
   export let allowEmpty: boolean = false;
+
+  function handleChange(event: Event) {
+    const select = event.target as HTMLSelectElement;
+
+    const CHANGE_EVENT_NAME = "khao-select-field-change";
+
+    const changeEvent = new CustomEvent(CHANGE_EVENT_NAME, {
+      detail: {
+        value: select.value,
+      },
+      bubbles: true,
+      composed: true,
+    });
+
+    select.dispatchEvent(changeEvent);
+  }
 </script>
 
 <div class="formfield">
   <label class="label" for={id}>{label}</label>
 
-  <select class="field" {id}>
+  <select class="field" {id} on:change={handleChange}>
     {#if allowEmpty}
       <option value=""></option>
     {/if}
 
     {#if options}
-    {#each JSON.parse(options) as { label, value }}
-      {#if selectedValue === value}
-        <option value={value} selected>{label}</option>
-      {:else}
-        <option value={value}>{label}</option>
-      {/if}
-    {/each}
+      {#each JSON.parse(options) as { label, value }}
+        {#if selectedValue === value}
+          <option {value} selected>{label}</option>
+        {:else}
+          <option {value}>{label}</option>
+        {/if}
+      {/each}
     {/if}
   </select>
 </div>
@@ -39,10 +54,8 @@
     background-size: var(--khao-sys-size-regular-6);
     background-position: right 2px bottom 0;
   }
- 
-   select.field {
+
+  select.field {
     padding-right: var(--khao-sys-size-regular-6);
-    
-   }
-  
+  }
 </style>