+++ src/components/misc/remoteControlled/types/RemoteControlEventDetail.ts
@@ -1,0 +1,6 @@
+import type { RemoteControlCommand } from "./RemoteControlCommand";
+
+export type RemoteControlEventDetail = {
+  channelId: string;
+  command: RemoteControlCommand;
+};