+++ src/components/misc/remoteControlled/types/Visibility.ts
@@ -1,0 +1,5 @@
+export const visibilityOptions = ["visible", "hidden"];
+
+export type Visibility = (typeof visibilityOptions)[number];
+
+export const Visibilitydefault: Visibility = "hidden";