+++ src/components/misc/remoteControlled/utils/dispatchRemoteControlEvent.ts
@@ -1,0 +1,23 @@
+import { remoteControlEventName } from "../types/Costants";
+import type { RemoteControlCommand } from "../types/RemoteControlCommand";
+import type { RemoteControlEventDetail } from "../types/RemoteControlEventDetail";
+
+const dispatchRemoteControlEvent = (
+  channelId: string,
+  command: RemoteControlCommand
+) => {
+  const payload: RemoteControlEventDetail = {
+    channelId: channelId,
+    command: command,
+  };
+
+  const remoteControlEvent = new CustomEvent(remoteControlEventName, {
+    detail: payload,
+    bubbles: true,
+    composed: true,
+  });
+
+  document.dispatchEvent(remoteControlEvent);
+};
+
+export default dispatchRemoteControlEvent;